reorganizes libs into directories to make documentation clearer
