+++ lib/athena/index.js
@@ -3,21 +3,19 @@
 const aws = require('aws-sdk');
 const Promise = require('bluebird');
 const _ = require('lodash');
-const proxy = require('proxy-agent');
 
 class AthenaQueryClient {
-  constructor ({ database, s3Bucket, region, httpProxy }) {
+  constructor ({ database, s3Bucket, region }) {
     this.database = database;
     this.s3Bucket = s3Bucket;
-
-    const athenaOptions = { region };
-    if (httpProxy) {
-      athenaOptions = { agent: proxy( httpProxy ) };
-    }
-    // configure athena
-    this.athena = new aws.Athena(athenaOptions);
+    this.athena = new aws.Athena({ region });
   }
 
+  /**
+   * Starts an Athena query execution
+   * 
+   * @param {String} sql 
+   */
   startQuery(sql) {
     return this.athena
       .startQueryExecution({
@@ -32,6 +30,12 @@ class AthenaQueryClient {
       .promise();
   }
 
+  /**
+   * Recursively polls Athena for results based on the query execution object. Will Resolve when query has succeeded, will reject if query fails.
+   * 
+   * @param {} queryInfo 
+   * @returns {Promise}
+   */
   pollForResults(queryInfo) {
     return this.athena
       .getQueryExecution(queryInfo)
