+++ lib/cloudsearch/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const AWS = require('aws-sdk');
-const clientConfig = require('./client-config');
+const clientConfig = require('../client-config');
 const _ = require('lodash');
-const promisify = require('./utils').promisify;
+const promisify = require('../utils').promisify;
 
 function createClient(config = {}) {
   config = _.merge(_.cloneDeep(clientConfig), config);
