+++ lib/dynamo/dynamo-helpers.js
