+++ lib/dynamo/index.js
@@ -3,9 +3,9 @@
 const AWS = require('aws-sdk');
 const helpers = require('./dynamo-helpers');
 const _ = require('lodash');
-const clientConfig = require('./client-config');
+const clientConfig = require('../client-config');
 const https = require('https');
-const promisify = require('./utils').promisify;
+const promisify = require('../utils').promisify;
 
 /**
  * create an http agent for the given configuration
