+++ lib/sns/index.js
@@ -2,7 +2,7 @@
 
 const AWS = require('aws-sdk');
 const Promise = require('bluebird');
-const clientConfig = require('./client-config');
+const clientConfig = require('../client-config');
 
 module.exports = function init({
   account,
