+++ lib/sqs/index.js
@@ -3,15 +3,13 @@
 const _ = require('lodash');
 const AWS = require('aws-sdk');
 const Promise = require('bluebird');
-const clientConfig = require('./client-config');
+const clientConfig = require('../client-config');
 const uuid = require('uuid/v4');
 const util = require('util');
 const zlib = require('zlib');
 const compress = util.promisify(zlib.gzip);
 const decompress = util.promisify(zlib.gunzip);
-const computeMessageSize = require('./utils').computeMessageSize;
-const promisify = require('./utils').promisify;
-const setHttpAgent = require('./utils').setHttpAgent;
+const {computeMessageSize, promisify, setHttpAgent} = require('../utils');
 
 // 256kb => b
 const MAX_MSG_SIZE = 262144;
