+++ test/unit/lib/dynamo-helpers.spec.js
@@ -2,7 +2,7 @@
 
 const chai = require('chai');
 const expect = chai.expect;
-const helpers = require('../../../lib/dynamo-helpers');
+const helpers = require('../../../lib/dynamo/dynamo-helpers');
 
 describe('Dynamo Helpers', function() {
   let testHash;
